Fix building unittests for arm
Before this commit building unittests was controlled with cmake option
RUN_TESTS. If this option was ON, shared library being built was not
operable because it lacked PIGPIO. RUN_TESTS option could be used to
build for platforms for which PIGPIO could not be built or did not work
correctly, so you could run unittests at least. However, when building
for arm, with the option ON you obtained non-functional driver shared
library, and without -- the tests that failed (because calls to
hardware from the tests were not mocked/faked).

This commit introduces several changes:
* unittests always use fakehardware, that replaces calls to the
hardware implementation compiled into tested rtuartscreader_static lib;
* driver shared library does not get built on the platforms that do not
support PIGPIO (presumably, all non-arms); static library uses dummy
version of hardware, that returns an error on any call;
* unittests won't run during build if target platform processor does
not match the host one;
* with the cmake option -DRTUARTSCREADER_BUILD_UNITTESTS=OFF building
unittests can be disabled
* with the cmake option -DRTUARTSCREADER_RUN_UNITTESTS=OFF running
unittests during build can be disabled